refactor: types.ts file move into folder component that uses it and change file name in API foder get.ts to index.ts
