+++ src/app/components/hompage/Category_Section/index.tsx
@@ -2,8 +2,8 @@
 
 import { useState } from "react";
 import { Tag, TagSkeleton } from "@/app/components/common/Tag";
-import { useCategories } from "@/API/hooks/useCategories";
-import { CategoryProps } from "@/API/tag-categories/types";
+import { useCategories } from "@/app/hooks/useCategories";
+import { CategoryProps } from "@/app/components/hompage/Category_Section/types";
 
 const Category = () => {
   const [selectedCategory, setSelectedCategory] = useState("");
