+++ src/app/components/hompage/Category_Section/types.ts
