+++ src/app/components/hompage/Header_Section/index.tsx
@@ -1,12 +1,13 @@
 import { IoBagOutline, IoNotificationsOutline } from "react-icons/io5";
 import { Search } from "@/app/components";
-import { useUserProfile } from "@/API/hooks/useUserProfile";
+import { useUserProfile } from "@/app/hooks/useUserProfile";
 import { SkeletonHeader } from "./SkeletonHeader";
 import { Button } from "@/components/ui/button";
 import Image from "next/image";
 import Link from "next/link";
+
 const Header = () => {
-  const { userProfile, isLoading, isError } = useUserProfile();
+  const { user, isLoading, isError } = useUserProfile();
 
   if (isError) {
     switch (isError) {
@@ -26,9 +27,9 @@ const Header = () => {
       <header>
         <div className="flex justify-between items-center mt-14 mx-6">
           <div className="flex flex-col justify-center gap-2">
-            {isLoading && userProfile ? (
+            {isLoading && user ? (
               <SkeletonHeader />
-            ) : !userProfile ? (
+            ) : !user ? (
               <div className="flex flex-col justify-center gap-2">
                 <p>Selamat Datang,</p>
                 <h4 className="font-bold text-xl font-jakarta">
@@ -49,14 +50,14 @@ const Header = () => {
               <>
                 <p>Selamat Datang,</p>
                 <h4 className="font-bold text-xl font-jakarta">
-                  {userProfile.firstname} {userProfile.lastname}
+                  {user.firstname} {user.lastname}
                 </h4>
               </>
             )}
           </div>
 
           <div className="flex justify-center items-center gap-3">
-            {!userProfile ? (
+            {!user ? (
               <Image src="/logo_toko.png" alt="logo" width={100} height={125} priority />
             ) : (
               <>
