+++ src/app/components/hompage/Header_Section/types.ts
