+++ src/app/components/hompage/ProductionCard/types.ts
@@ -1,10 +1,3 @@
-export interface PromoProps {
-    id: number;
-    name: string;
-    photo_url: string;
-    promo_special: number;
-}
-
 export interface ProductionProps {
     id: number;
     name: string;
