+++ src/app/components/hompage/Production_Section/types.ts
@@ -1,0 +1,8 @@
+export interface ProductionProps {
+    id: number;
+    name: string;
+    photo_url: string;
+    description_list: string[];
+    category: string;
+    created_at: string;
+}
