+++ src/app/components/hompage/Promo_Section/types.ts
@@ -1,0 +1,6 @@
+export interface PromoProps {
+    id: number;
+    name: string;
+    photo_url: string;
+    promo_special: number;
+}
